pyproject.toml
[project]
name = "my-chess-style"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "berserk>=0.13.2",
  "celery[redis]>=5.5.0",
  "chess-com>=3.11.1",
  "django>=5.2",
  "django-ninja>=1.4.0",
  "fastparquet>=2024.11.0",
  "huggingface-hub>=0.31.2",
  "langchain-ollama>=0.3.6",
  "matplotlib>=3.10.1",
  "numpy>=2.2.4",
  "pandas>=2.2.3",
  "psycopg2>=2.9.10",
  "pyarrow>=20.0.0",
  "python-dotenv>=1.1.0",
  "requests>=2.32.3",
  "scikit-learn>=1.6.1",
  "scipy>=1.15.2",
  "stockfish",
]

[dependency-groups]
commit = ["pre-commit>=4.2.0"]
data_training = [
  "jupyterlab>=4.4.0",
  "mglearn>=0.2.0",
  "notebook>=7.4.0",
  "seaborn>=0.13.2",
  "voila>=0.5.8",
  "matplotlib>=3.10.1",
  "numpy>=2.2.4",
  "pandas>=2.2.3",
  "scikit-learn>=1.6.1",
  "scipy>=1.15.2",
  "kagglehub[pandas-datasets]>=0.3.11",
  "imbalanced-learn>=0.13.0",
]
dev = [
  "celery-types>=0.23.0",
  "chess-com>=3.11.1",
  "coverage>=7.9.2",
  "django-stubs>=5.1.3",
  "django-types>=0.20.0",
  "flower>=2.0.1",
  "ipython>=9.1.0",
  "pytest>=8.3.5",
  "pytest-cov>=6.2.1",
  "ruff>=0.11.3",
  "types-requests>=2.32.0.20250328",
  "watchdog>=6.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.bandit.assert_used]
skips = ['*_test.py', '*/test_*.py']

[tool.basedpyright]
venvPath = "./"
venv = ".venv"

[tool.uv.sources]
stockfish = { git = "https://github.com/py-stockfish/stockfish", rev = "master" }

[tool.coverage.run]
source = ["style_predictor"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
